requirement.txt
accelerate==0.24.1
beartype==0.16.4
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
denoising-diffusion-pytorch==1.9.4
easydict==1.13
einops==0.7.0
ema-pytorch==0.3.0
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
huggingface-hub==0.19.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.32.0
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.3
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
protobuf==4.25.0
psutil==5.9.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytorch-fid==0.3.0
PyWavelets==1.4.1
PyYAML==6.0.1
requests @ file:///croot/requests_1690400202158/work
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tifffile==2023.7.10
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.66.1
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
urllib3 @ file:///croot/urllib3_1698257533958/work
zipp==3.17.0
